+++ src/model/profiles.rs
@@ -309,6 +309,7 @@ mod tests {
             arch: "x86_64".into(),
             host: "laptop".into(),
             family: "debian".into(),
+            vars: Default::default(),
         }
     }
 
@@ -423,6 +424,7 @@ mod active_tests {
             arch: "x86_64".into(),
             host: host.into(),
             family: "debian".into(),
+            vars: Default::default(),
         }
     }